Safari-611.1.10/LayoutTests/inspector/sampling-profiler/basic-expected.txt


== Running test suite: ScriptProfiler.Samples.Basic
-- Running test case: Sampling Profiler basic
PASS: Should have seen stacktrace:
[
  {
    "name": "foo"
  },
  {
    "name": "runFor"
  }
]
